refactor: OrderTable -> Order 단방향 설정, TableValidator 제거
