+++ src/main/java/kitchenpos/table/domain/OrderTable.java
@@ -1,10 +1,16 @@
 package kitchenpos.table.domain;
 
+import kitchenpos.order.domain.Order;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToMany;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 
 @Entity
@@ -21,6 +27,10 @@ public class OrderTable {
 
     private boolean empty;
 
+    @OneToMany
+	@JoinColumn(name = "orderTableId")
+	private List<Order> orders = new ArrayList<>();
+
     protected OrderTable() {
     }
 
@@ -28,6 +38,14 @@ public class OrderTable {
         validateNumberOfGuests(numberOfGuests);
         this.numberOfGuests = numberOfGuests;
         this.empty = empty;
+        this.orders = new ArrayList<>();
+    }
+
+    private OrderTable(int numberOfGuests, boolean empty, List<Order> orders) {
+        validateNumberOfGuests(numberOfGuests);
+        this.numberOfGuests = numberOfGuests;
+        this.empty = empty;
+        this.orders = orders;
     }
 
     private void validateNumberOfGuests(int numberOfGuests) {
@@ -40,11 +58,16 @@ public class OrderTable {
         return new OrderTable(numberOfGuests, empty);
     }
 
+    public static OrderTable ofOrder(int numberOfGuests, boolean empty, List<Order> orders) {
+        return new OrderTable(numberOfGuests, empty, orders);
+    }
+
     public void changeEmpty(boolean empty) {
         if (isGrouped()) {
             throw new IllegalArgumentException("단체 지정 된 테이블은 상태를 변경할 수 없습니다.");
         }
 
+        validateOrderStatus();
         this.empty = empty;
     }
 
@@ -54,6 +77,7 @@ public class OrderTable {
     }
 
     public void ungroup() {
+        validateOrderStatus();
         this.tableGroupId = null;
     }
 
@@ -86,5 +110,14 @@ public class OrderTable {
 
     private void grouped() {
         this.empty = false;
+    }
+
+    private void validateOrderStatus() {
+        boolean isNotCooking = orders
+                .stream()
+                .anyMatch(Order::isNotCooking);
+        if (isNotCooking) {
+            throw new IllegalArgumentException("주문 테이블의 상태가 조리이거나, 식사이면 변경할 수 없습니다.");
+        }
     }
 }
