--- src/test/java/kitchenpos/table/application/TableValidatorTest.java
@@ -1,77 +1,0 @@
-package kitchenpos.table.application;
-
-import kitchenpos.common.fixtrue.MenuFixture;
-import kitchenpos.common.fixtrue.MenuGroupFixture;
-import kitchenpos.common.fixtrue.MenuProductFixture;
-import kitchenpos.common.fixtrue.OrderTableFixture;
-import kitchenpos.common.fixtrue.ProductFixture;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuProducts;
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderLineItems;
-import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.product.domain.Product;
-import kitchenpos.table.domain.OrderTable;
-import org.assertj.core.api.Assertions;
-import org.assertj.core.api.ThrowableAssert;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import java.math.BigDecimal;
-import java.util.Collections;
-
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
-
-@ExtendWith(MockitoExtension.class)
-class TableValidatorTest {
-
-    @Mock
-    OrderRepository orderRepository;
-
-    @InjectMocks
-    TableValidator tableValidator;
-
-    Menu 후라이드_후라이드;
-    OrderTable 주문_테이블;
-    OrderLineItems 주문_항목들;
-
-    @BeforeEach
-    void setUp() {
-        Product 후라이드치킨 = ProductFixture.of("후라이드치킨", BigDecimal.valueOf(16000));
-        MenuGroup 두마리치킨 = MenuGroupFixture.from("두마리치킨");
-        MenuProduct 후라이드_메뉴_상품 = MenuProductFixture.of(후라이드치킨.getId(), 2);
-        후라이드_후라이드 = MenuFixture.of(
-                "후라이드+후라이드",
-                BigDecimal.valueOf(31000),
-                두마리치킨,
-                MenuProducts.from(Collections.singletonList(후라이드_메뉴_상품)));
-        주문_테이블 = OrderTableFixture.of(4, false);
-        주문_항목들 = OrderLineItems.from(Collections.singletonList(OrderLineItem.of(후라이드_후라이드, 1L)));
-    }
-
-    @Test
-    void 주문상태가_조리이거나_식사이면_변경할_수_없다() {
-        // given
-        Order 주문 = Order.of(1L, 주문_항목들);
-        주문.changeOrderStatus(OrderStatus.MEAL);
-        given(orderRepository.findByOrderTableId(any())).willReturn(주문);
-
-        // when
-        ThrowableAssert.ThrowingCallable throwingCallable = () -> tableValidator.validateOrderStatus(주문_테이블);
-
-        // then
-        Assertions.assertThatExceptionOfType(IllegalArgumentException.class)
-                .isThrownBy(throwingCallable)
-                .withMessage("주문 테이블의 상태가 조리이거나, 식사이면 변경할 수 없습니다.");
-    }
-
-}
